[query] when testing, ensure our thread has an event loop
`asyncio.get_event_loop` refuses to start an event loop for you if you are not the
main thread. This is normally sensible: you should only have one event loop on one
thread and all async work should be sent to that thread's event loop. However,
during testing, it appears that pytest sometimes places us on a thread named
"Dummy-1". I do not understand why. I cannot find documentation about this. Anyway,
an easy fix is to create an event loop in whatever thread is running the tests.