[security] change key age to 5 days & only consider non-system-managed keys
A job running for five days *could* cause an issue here, so the operator should still be cognizant of that. The
move to access tokens will eliminate that concern anyway.

We should not consider system-managed keys because those are managed by Google and not even visible in the
UI. We are not obligated to rotate them (nor are we able, afaik).